requirements.txt
langchain
pandas
google-generativeai>=0.8.2
tabulate
gradio
pymongo